+++ internal/cli/agent/agent.go
@@ -13,7 +13,15 @@ func SetAPIClient(client *client.Client) {
 }
 
 var AgentCmd = &cobra.Command{
-	Use: "agent",
+	Use:   "agent",
+	Short: "Commands for managing agents",
+	Long:  `Commands for managing agents.`,
+	Args:  cobra.ArbitraryArgs,
+	Example: `arctl agent list
+arctl agent show dice
+arctl agent publish ./my-agent
+arctl agent remove dice
+arctl agent run ./my-agent`,
 }
 
 func init() {
@@ -22,4 +30,12 @@ func init() {
 	AgentCmd.AddCommand(InitCmd)
 	AgentCmd.AddCommand(BuildCmd)
 	AgentCmd.AddCommand(RunCmd)
+	AgentCmd.AddCommand(AddSkillCmd)
+	AgentCmd.AddCommand(AddMcpCmd)
+	AgentCmd.AddCommand(PublishCmd)
+	AgentCmd.AddCommand(RemoveCmd)
+	AgentCmd.AddCommand(DeployCmd)
+	AgentCmd.AddCommand(UnpublishCmd)
+	AgentCmd.AddCommand(ListCmd)
+	AgentCmd.AddCommand(ShowCmd)
 }
