+++ internal/cli/agent/deploy.go
@@ -1,0 +1,21 @@
+package agent
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+var DeployCmd = &cobra.Command{
+	Use:   "deploy [name] [args...]",
+	Short: "Deploy an agent",
+	Long:  `Deploy an agent. Use flags for non-interactive setup or run without flags to open the wizard.`,
+	Args:  cobra.ArbitraryArgs,
+	RunE:  runDeploy,
+}
+
+func runDeploy(cmd *cobra.Command, args []string) error {
+	// Not implemented yet
+	fmt.Println("Not implemented yet")
+	return nil
+}
