+++ internal/cli/mcp/mcp.go
@@ -13,7 +13,15 @@ func SetAPIClient(client *client.Client) {
 }
 
 var McpCmd = &cobra.Command{
-	Use: "mcp",
+	Use:   "mcp",
+	Short: "Commands for managing MCP servers",
+	Long:  `Commands for managing MCP servers.`,
+	Args:  cobra.ArbitraryArgs,
+	Example: `arctl mcp list
+arctl mcp show my-mcp-server
+arctl mcp publish ./my-mcp-server
+arctl mcp deploy my-mcp-server
+arctl mcp remove my-mcp-server`,
 }
 
 func init() {
@@ -23,4 +31,9 @@ func init() {
 	McpCmd.AddCommand(BuildCmd)
 	McpCmd.AddCommand(AddToolCmd)
 	McpCmd.AddCommand(PublishCmd)
+	McpCmd.AddCommand(DeployCmd)
+	McpCmd.AddCommand(RemoveCmd)
+	McpCmd.AddCommand(ListCmd)
+	McpCmd.AddCommand(RunCmd)
+	McpCmd.AddCommand(ShowCmd)
 }
