+++ internal/cli/skill/utils.go
@@ -1,0 +1,21 @@
+package skill
+
+import "strings"
+
+// sanitizeRepoName converts a skill name to a docker-friendly repo name
+func sanitizeRepoName(name string) string {
+	n := strings.TrimSpace(strings.ToLower(name))
+	// replace any non-alphanum or separator with dash
+	// also convert path separators to dashes
+	replacer := strings.NewReplacer("/", "-", "\\", "-", " ", "-")
+	n = replacer.Replace(n)
+	// collapse consecutive dashes
+	for strings.Contains(n, "--") {
+		n = strings.ReplaceAll(n, "--", "-")
+	}
+	n = strings.Trim(n, "-")
+	if n == "" {
+		n = "skill"
+	}
+	return n
+}
